chore(packaging): move to src/ layout and single pyproject.toml
Consolidates packaging metadata into PEP 621 pyproject.toml with hatchling
as the build backend. Deletes setup.py, setup.cfg, and requirements.txt
which previously disagreed on the spaCy lower bound and the version string.

- Move negspacy/ -> src/negspacy/ (src/ layout)
- Populate src/negspacy/__init__.py with __version__ = "1.0.5" for
  [tool.hatch.version]
- Add src/negspacy/py.typed (PEP 561 marker; hints already exist in
  negation.py)
- Split negspacy/test.py into tests/test_negation.py and
  tests/test_termsets.py (no assertion changes)
- Add tests/conftest.py with session-scoped en_core_web_sm fixture that
  restores the pipeline between tests
- Quarantine negspacy/test_scispacy_dep.py as tests/_scispacy_dep.py
  (underscore prefix keeps pytest from collecting the broken imports;
  PR #6 will un-break or delete)
- Add pytest and ruff configuration to pyproject.toml
- Configure dev extras (pytest, pytest-cov, ruff, pre-commit)

Verified with `pip install -e ".[dev]" && pytest`: 7 passed, 93% coverage.
Wheel build via `python -m build` produces a clean artifact containing
only the package code, py.typed, and the license.
