--- setup.cfg
@@ -1,32 +1,0 @@
-[metadata]
-name = negspacy
-version = 1.0.5
-description = spaCy pipeline object for negating concepts in text
-url = https://github.com/jenojp/negspacy
-author = Jeno Pizarro
-author_email = jenopizzaro@gmail.com
-license = MIT
-long_description = file: README.md
-long_description_content_type = text/markdown
-classifiers =
-    Development Status :: 4 - Beta
-    Environment :: Console
-    Intended Audience :: Developers
-    Intended Audience :: Science/Research
-    License :: OSI Approved :: MIT License
-    Operating System :: POSIX :: Linux
-    Operating System :: MacOS :: MacOS X
-    Operating System :: Microsoft :: Windows
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Topic :: Scientific/Engineering
-
-[options]
-python_requires = >=3.6
-install_requires =
-    spacy>=3.0.2,<4.0.0
