+++ src/negspacy/negation.py
