+++ src/negspacy/termsets.py
