+++ tests/test_negation.py
@@ -1,10 +1,7 @@
-import pytest
-import spacy
-import copy
-import negspacy.negation
+import negspacy.negation  # noqa: F401
 from negspacy.termsets import termset
-from spacy.pipeline import EntityRuler
 from spacy.language import Language
+from spacy.pipeline import EntityRuler
 
 
 def build_docs():
@@ -39,8 +36,7 @@ def build_docs():
     return docs
 
 
-def test():
-    nlp = spacy.load("en_core_web_sm")
+def test(nlp):
     nlp.add_pipe("negex", last=True)
     docs = build_docs()
     for d in docs:
@@ -50,8 +46,7 @@ def test():
             assert (e.text, e._.negex) == d[1][i]
 
 
-def test_en():
-    nlp = spacy.load("en_core_web_sm")
+def test_en(nlp):
     ts = termset("en")
     nlp.add_pipe("negex", config={"neg_termset": ts.get_patterns()}, last=True)
     docs = build_docs()
@@ -62,8 +57,7 @@ def test_en():
             assert (e.text, e._.negex) == d[1][i]
 
 
-def test_own_terminology():
-    nlp = spacy.load("en_core_web_sm")
+def test_own_terminology(nlp):
     nlp.add_pipe(
         "negex",
         config={
@@ -80,67 +74,11 @@ def test_own_terminology():
     assert doc.ents[1]._.negex == False
 
 
-def test_issue7():
-    nlp = spacy.load("en_core_web_sm")
+def test_issue7(nlp):
     nlp.add_pipe("negex", last=True)
     ruler = EntityRuler(nlp)
     patterns = [{"label": "SOFTWARE", "pattern": "spacy"}]
     doc = nlp("fgfgdghgdh")
-
-
-def test_get_patterns():
-    # nlp = spacy.load("en_core_web_sm")
-    # negex = Negex(nlp)
-    # patterns = negex.get_patterns()
-    ts = termset("en")
-    patterns = ts.get_patterns()
-    assert type(patterns) == dict
-    assert len(patterns) == 4
-
-
-def test_add_remove_patterns():
-    ts = termset("en_clinical")
-    patterns = copy.deepcopy(ts.get_patterns())
-    ts.add_patterns(
-        {
-            "pseudo_negations": ["my favorite pattern"],
-            "termination": ["these are", "great patterns", "but"],
-            "preceding_negations": ["wow a negation"],
-            "following_negations": ["extra negation"],
-        }
-    )
-    patterns_after = ts.get_patterns()
-
-    assert len(patterns_after["pseudo_negations"]) - 1 == len(
-        patterns["pseudo_negations"]
-    )
-    assert len(patterns_after["termination"]) - 2 == len(patterns["termination"])
-    assert len(patterns_after["preceding_negations"]) - 1 == len(
-        patterns["preceding_negations"]
-    )
-    assert len(patterns_after["following_negations"]) - 1 == len(
-        patterns["following_negations"]
-    )
-
-    ts.remove_patterns(
-        {
-            "termination": ["these are", "great patterns"],
-            "pseudo_negations": ["my favorite pattern"],
-            "preceding_negations": ["denied", "wow a negation"],
-            "following_negations": ["unlikely", "extra negation"],
-        }
-    )
-    patterns_after = ts.get_patterns()
-    assert len(patterns_after["termination"]) == len(patterns["termination"])
-    assert (
-        len(patterns_after["following_negations"])
-        == len(patterns["following_negations"]) - 1
-    )
-    assert (
-        len(patterns_after["preceding_negations"])
-        == len(patterns["preceding_negations"]) - 1
-    )
-    assert len(patterns_after["pseudo_negations"]) == len(patterns["pseudo_negations"])
 
 
 def ents_to_spans(doc):
@@ -166,8 +104,7 @@ def convert_ents_to_spans(doc):
     return doc
 
 
-def test_spans():
-    nlp = spacy.load("en_core_web_sm")
+def test_spans(nlp):
     nlp.add_pipe("ents_to_spans", last=True)
     nlp.add_pipe("negex", last=True, config={"span_keys": ["ent_spans"]})
 
@@ -177,13 +114,3 @@ def test_spans():
         for i, e in enumerate(doc.spans["ent_spans"]):
             print(e.text, e._.negex)
             assert (e.text, e._.negex) == d[1][i]
-
-
-if __name__ == "__main__":
-    test()
-    test_en()
-    test_own_terminology()
-    test_get_patterns()
-    test_issue7()
-    test_add_remove_patterns()
-    test_spans()
